Boeffla ramfs: Special treatment for Lineage 14.1 cpu permissions
We noticed in Lineage, the battery-current-limit (BCL) driver sometimes
seems to reset permissions of the online sysfs file for CPU4. Hence
we cannot bring this CPU core properly up and down in bccontroller anymore.
Permissions for CPU cores 4-7 are now corrected every time before we
access the online file in sysfs.