Squash latest LineageOS commits up to 26.07.2017, 13:00 CET
ANDROID: mnt: Fix next_descendent

next_descendent did not properly handle the case
where the initial mount had no slaves. In this case,
we would look for the next slave, but since don't
have a master, the check for wrapping around to the
start of the list will always fail. Instead, we check
for this case, and ensure that we end the iteration
when we come back to the root.

Signed-off-by: Daniel Rosenberg <drosen@google.com>
Bug: 62094374
Change-Id: I43dfcee041aa3730cb4b9a1161418974ef84812e

ANDROID: sdcardfs: Remove unnecessary lock

The mmap_sem lock does not appear to be protecting
anything, and has been removed in Samsung's more
recent versions of sdcardfs.

Signed-off-by: Daniel Rosenberg <drosen@google.com>
Change-Id: I76ff3e33002716b8384fc8be368028ed63dffe4e
Bug: 63785372

Add: CONFIG_SECURITY_PERF_EVENTS_RESTRICT=y

to android-base.cfg

The kernel.perf_event_paranoid sysctl is set to 3 by default.
No unprivileged use of the perf_event_open syscall will be
permitted unless it is changed.

Bug: 29054680
Change-Id: Ie7512259150e146d8e382dc64d40e8faaa438917

oneplus2: defconfig: Enable CONFIG_SECURITY_PERF_EVENTS_RESTRICT

* This flag was introduced as a part of CVE-2016-3843

Change-Id: Idfc2358d675db99ff05d533f37324a1dca2437cd

memcg: add permission check

Use the 'allow_attach' handler for the 'mem' cgroup to allow
non-root processes to add arbitrary processes to a 'mem' cgroup
if it has the CAP_SYS_NICE capability set.

Bug: 18260435
Signed-off-by: Rom Lemarchand <romlem@android.com>
Change-Id: If7d37bf90c1544024c4db53351adba6a64966250
Signed-off-by: Corinna Vinschen <xda@vinschen.de>

Revert "memcg: add permission check"

This is an old hook which is reported not working at all. Correct commit
by CAF has already been merged:
https://github.com/LineageOS/android_kernel_oneplus_msm8994/commit/d68f06d49150d08a952f466791d05b81e42ab6f5

And the commit itself is wrong enough, it will cause a build
failure as the functions have same name.

This reverts commit b5970d1bb0bc65c2d24f4e5158b97a84d635b6d7.

Change-Id: I865c442f0d8e3fe8a169a622c388d1816524e8d6